Divide Header.js into many component files
