+++ src/Components/Main/Header/HeaderLinkGroup.js
@@ -1,0 +1,8 @@
+import styled from "styled-components"
+
+const HeaderLinkGroup = styled.div`
+  display: flex;
+  align-items: center;
+`
+
+export default HeaderLinkGroup
