+++ src/Components/Main/Header/HeaderStartButton.js
@@ -1,0 +1,17 @@
+import styled from "styled-components"
+
+const HeaderStartButton = styled.button`
+  background-color: black;
+  color: white;
+  font-size: 14px;
+  width: 105.5px;
+  height: 38px;
+  border: 0;
+  box-sizing: border-box;
+  border-radius: 4px;
+  &:hover {
+    cursor: pointer;
+  }
+`
+
+export default HeaderStartButton
